paterson.c: replace element_to_mpz().
It seems element_from_hash() is what is needed.

Also changed t3 to be a element of Zr.

The element_to_mpz() function is meant to convert a PBC integer to
a GMP integer. There are no guarantees on its effects on anything
besides integers, though it seems for field extensions it returns
the constant term as I once thought that might be useful.